requirements.txt
absl-py==2.2.2
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
decorator==5.2.1
Django==5.2
executing==2.2.0
filelock==3.18.0
filterpy==1.4.5
firebase-admin==6.8.0
Flask==3.1.0
flask-ngrok==0.0.25
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gast==0.4.0
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
httplib2==0.22.0
huggingface-hub==0.31.2
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
keras==3.10.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.9
networkx==3.4.2
nlpaug==1.1.11
numpy==2.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
svgwrite==1.4.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
Tree==0.2.4
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.140
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wget==3.2
windows-curses==2.4.1
wrapt==1.14.1
xgboost==3.0.0
